Remember specified handler if preferred is undefined
In the situation when dired-launch-extensions-map does not specify a
user-specified handler, offer to "remember" the user-specified
handler, if specified via the prompt. The selection only persists for
the current session.

We provide dired-launch-extensions-map-pop to allow the user to ditch
the preferred handler, for the current session, if it is no longer
desirable.